Update background_base.cc
Fixed a bug in the "NumericalDerivatives()" function. The vector gradients of the fields are initially stored in _spdata_tmp (see "DirectionalDerivative()") before they are copied to _spdata at the end of the function. Prior to copying, the matrices storing the gradients should be transposed in order to match standard notation grad(V)_ij = dV_i / dx_j.  However, the gradients in _spdata were being mistakenly transposed, instead of the ones in _spdata_tmp.