Added New Diffusion Class
Added a DiffusionMomentumPowerLaw class. It is not a special case of the DiffusionRigidityMagneticFieldPowerLaw (for power_law_B = 0) because it uses a reference value for the diffusion coefficient directly instead of the parallel mean free path which is then multiplied by velocity and divided by a factor of 3. Therefore, the diffusion coefficients do not depend on particle velocity outside of the specified momentum power law.

Additionally, an override version of GetMuDerivative() was added to all those classes for which it is trivially zero for efficiency.