Updates to BackgroundBase and BackgroundSolarWind
BackgroundBase now has an initial time, "t0", attribute which it unpacks from the data container during the "SetupBackground()"  routine. It then uses it in the "GetFields()" method. This change was propagated though the codes in the benchmarks and runs folders.

BackgroundSolarWind and BackgroundSolarWindTermShock now include an option for a wavy current sheet with solar cycle (time) dependence. The separation between sectors accounts for the, possibly non-uniform, radial solar wind speed through the "TimeLag()" function.