Numerical Derivatives and Other Updates
In this major update, the numerical derivatives in both Background and Diffusion objects have been standardized to use central finite differences (2nd order) and use the particle's Larmor radius as the increment distance. This is acceptable since the derivative attributes of SpatialData objects are only used in transport models for which the scales of variation of the background are assumed to be much larger than the Larmor radius. In order to do this, the momentum magnitude has to be passed into "GetFields()". This change was propagated through the relevant files/functions. "ConvertMomentum()" was standarized as always returning momentum in (p, mu, phi) coordinates.

Features were added to the BackgroundSolarWind and derived classes related to the solar cycle and extent of the HCS, including a parameter to control the relative thickness of the peaks and troughs of the cycle.

In addition, an InitialTime condition was introduced for Trajectory objects to sync their background with respect to the "t0" in the Background classes in time dependent runs. This is the initial time for forward runs, and the final time for backward runs.

SpatialData now has a function to compute the time derivative of the magnetic unit vector called "dbhatdt()". This change was used in "Slopes()" for the TrajectoryFocused class and "ModifiedFields()" in the TrajectoryGuiding class.

A bugs were also fixed, such as a missing "double t0" attribute in the BackgroundBase class or a stray "virtual" in "ConvertMomentum()" within background_base.hh. <algorithm> had to be (#)included into "simulation.hh" in order to use "std::max_element()". <iostream> was removed from packages that did not use it.