Bug Fix to Solar Wind Background
Fixed the Smith & Bieber correction to the Parker Spiral. Previously, it used "r_mns" ("r - r_ref") to compute the correction factor, when the formula technically uses "r". This change is negligible in a majority of the domain, since by 1 au "r" >> "r_ref". However, it does make the field non-radial at the source surface. Therefore, this bug fix was made for correctness.