Created new background server class
A new class was created to be the common ancestor of all background interfaces with grids on distributed memory (server - worker model). This new class, called BackgroundServer, was located in the class family as a child of BackgroundBase and the parent of BackgroundCartesian. This change was made in order to distinguish this branch of Background classes from those for which the grid is defined on shared memory (future development).

The Makefile.am in the benchmarks folder was also updated to reflect this update.