Updated Makefiles and benchmark codes
`Makefile.am` in the "runs" and "benchmark" directories were updated to include missing header file dependencies in accordance with the agreed upon standard of listing dependencies.

The documentation was also expanded with the "Examples" chapter, describing a few applications of the software. Some minor edits/updates were also made.

`gcr_modulation_example.cc` was missing a required parameter when initializing the `DistributionTimeUniform()` object, which was causing runtime errors.

A few benchmark tests were updated to adhere to the new `MPI_Config` implementation.

A comment was changed in `TimeBoundaryProximityCheck` to better reflect the authorship and date of last update.

Finally, `DISTRO_KINETIC_ENERGY_POWER_LAW_TYPE` was set to `0` by default since one of the benchmark codes requires it for correct output.