Added features and name changes (#44)
* Updated background and new diffusion

Updated `BackgroundSolarWind` to include a time-dependent magnetic field intensity according to the solar cycle.

Also added a new diffusion module called `DiffusionPotgieterEtAl2015` for galactic cosmic ray electrons. It could have been derived from `DiffusionStraussEtAl2013` but I decided to keep it separate since it is still under development and it might change.

* Updated background class names

In preparation for the introduction "grid" background classes on shared memory, both `BackgroundCartesian` and `BackgroundBATL` have been renamed to `BackgroundServerCartesian` and `BackgroundServerBATL`. This emphasizes their distributed memory quality and the necessity for a "server" process to feed the workers. This change was propagated to the relevant benchmark tests.

**NOTE:** `background_batl.*` and `background_cartesian.*` should be deleted from `/src`. These files are not replaced by this updated due to the name change.