Update configure.ac (#47)
* Update configure.ac

Removed edits to `PKG_CONFIG_PATH` since they were redundant. If path was found then there is no need to reset `PKG_CONFIG_PATH`. Also, added the same check for GSL pkgconfig folder as for MPI.

* Update README.md

* Update diffusion_other.cc

Removed factor of 300 in rigidity calculations of diffusion coefficients.

* Update definitions.hh

Added an `Average` function to compute the arithmetic or geometric average of the values in an array.

* Updated gradUvec in solar wind backgrounds

Added an analytic calculation for `_spdata.gradUvec` in `BackgroundSolarWind` and `BackgroundSolarWindTermShock` background modules. The documentation was edited to reflect this change.

* Update background_solarwind_termshock.cc

Fixed bug. `posprime` should be `GeoVector` type, NOT `double`.

---------

Co-authored-by: Vladimir Florinski <79867224+vflorins@users.noreply.github.com>