Merged two diffusion classes
`DiffusionFlowPowerLaw` and `DiffusionMomentumPowerLaw` classes merged into single `DiffusionFlowMomentumPowerLaw` class. The individual flow velocity and momentum classes are deleted and replaced by single diffusion class which accounts for both the flow velocity and momentum dependence of the diffusion coefficient. The behavior of the single dependence of the diffusion may be created by making the other power law index equal to zero.