Merged two diffusion classes (#52)
* Merged two diffusion classes

`DiffusionFlowPowerLaw` and `DiffusionMomentumPowerLaw` classes merged into single `DiffusionFlowMomentumPowerLaw` class. The individual flow velocity and momentum classes are deleted and replaced by single diffusion class which accounts for both the flow velocity and momentum dependence of the diffusion coefficient. The behavior of the single dependence of the diffusion may be created by making the other power law index equal to zero.

* Update Documentation

Updated documentation to include the new `DiffusionFlowMomentumPowerLaw` class. The obsolete `DiffusionFlowPowerLaw` and `DiffusionMomentumPowerLaw` classes where removed from the text.

* Update to diffusion modules

Cleaned up `DiffusionStraussEtAl2013` and `DiffusionPotgieterEtAl2015` to better match the diffusion models in the publications they reference. Added a new diffusion module, `DiffusionEmpiricalSOQLTandUNLT`, which is a parametrized empirical fit to SOQLT (parallel) and UNLT (perpendicular) diffusion using a bent power law (injection + inertial range) for magnetic turbulence. Uploaded an updated documentation, consistent with these changes.

* Update background_solarwind.cc

Improved heliospheric current sheet calculation. The expression is not valid for ANY tilt angle, not just small tilt angles. The documentation will be updated in a different commit.

* Update documentation

Update documentation to include the new current sheet model in `BackgroundSolarWind`.

* Update README.md

* Upload citation file

* Update README.md

* Update BackgroundSmoothShock

There were typos in the analytical computation of `BackgroundSmoothShock::EvaluateBackgroundDerivatives`. These have been fixed. Also, the default derivative computation method is now analytical for this class, and the default shock profile is smooth.

* Improved BackgroundSmoothShock

The calculation of the distance to the shock was made more efficient by precomputing a dot product. The documentation regarding this background had a typo and it was corrected.

* Updated comments

* Added analytic derivative computations

Added analytic `DirectionalDerivative` computations for `DiffusionKineticEnergyRadialDistancePowerLaw` and `DiffusionRigidityMagneticFieldPowerLaw`, which follow the same pattern as in `DiffusionFlowMomentumPowerLaw`. The temporal directional derivative is now correctly handled in all three classes.

* Undo accidental changes

The previous commit accidentally erased some changes recently made in this branch. This commit reverts the code back to the correct version.

* Delete diffusion_other.cc

* Delete diffusion_other.hh

* Relocate files

Previous upload placed files in the wrong directory. This error has been fixed.

* Update documentation

Update documentation

* Update diffusion_other.hh

Added maximum radial extent for perpendicular coefficient reduction in `DiffusionEmpiricalSOQLTandUNLT`.

* Update diffusion_other.cc

Added maximum radial extent for perpendicular coefficient reduction in `DiffusionEmpiricalSOQLTandUNLT`.

* Update diffusion_other.hh

Changed attribute name in `DiffusionEmpiricalSOQLTandUNLT`.

* Update Documentation

Minor update to variable names.

* Update background_solarwind.cc

Added time-lagged solar cycle phase as an indicator variable.

* Updated BackgroundSmoothShock

The class `BackgroundSmoothShock` now has a variable `dmax` which changes linearly between a fraction of the shock width near the shock to `dmax0` far away from the shock.

Also, a new distribution class was added, `DistroPositionMomentumUniform` for applications in which the phase-space distribution is desired (e.g. shock acceleration).

* Update distribution_other.cc

Corrected a comment.

* Update distribution_other.hh

Corrected comment.

* Update background_solarwind.cc

Added a pre-compiler guard for the use of `W0_sw` in computing `Br0`.

* Update background_solarwind.cc

Switched the sign of `t_lag` to have a more natural direction.

---------

Co-authored-by: Juan G Alonso Guzman <33181561+jgaloguz@users.noreply.github.com>