Update distribution_templated.cc (#54)
The previous binning routine incorrectly assigned values just below the lower bin to the first bin due to integer rounding of floating points within (-1,0) to 0. This new version prevents that error by first checking if the value is below the lower bound, and only computing its bin if the value above the lower bound.