Update benchmark files
The output of the benchmark codes has been changed to the working directory, instead of `benchmarks/output_data` for better compatibility with other filesystem routines. The `run_tests.sh` script within the benchmarks subdirectory has been updated to create and utilize the same `output_data` subfolder so the behavior is the same as before.