Updated output for benchmark codes (#59)
* Update benchmark files

The output of the benchmark codes has been changed to the working directory, instead of `benchmarks/output_data` for better compatibility with other filesystem routines. The `run_tests.sh` script within the benchmarks subdirectory has been updated to create and utilize the same `output_data` subfolder so the behavior is the same as before.

* Update script to run tests

Added a command to return to the `benchmarks` subdirectory after running a code within the `output_data` subfolder so that multiple tests can be run sequentially.