Update background_solarwind_termshock.cc (#61)
* Update background_solarwind_termshock.cc

Incorporated a couple of bug fixes in functions relevant to derivative computation that were already discovered and fixed in a forked repository.

* Removed unnecessary parameter

Removed `r0_shock` from `BackgroundShock`, and hence also `BackgroundSmoothShock`, since the base class parameter `r0` was not being used and could perform the same function. The documentation was also updated to reflect this change.