Update to benchmarks and treatment of GSL dependency
The following issues have been fixed:
- The USE_GSL flag was not set in the configuration stage.
- The USE_GSL flag was not visible by random.hh because this file didn't import config.h.
- Makefile.am in the benchmarks folder did not list the main files, causing the linking stage to fail when trying to compile these codes.

Besides these issues, the todo comment relating to the particle specie has been removed from all benchmark tests because that issue was resolved. Also, the config.h dependency in physics.hh was changed from angled brackets to double quotes for consistency with the rest of the code.