Added draft of fields revision 'fields2', enabled for testing purposes.
Fields2 (Fields implementation in directory fields2) has a streamlined constructor and copy constructor, and preserves contiguous memory of Fields1. It also features a reduction in the amount of generated code, and an elegant, very fast reduction to a pure array of double. Species implementation also has improvements that reduce redundancies present in Fields1.