Changes to the fields implementation to account for the new policy, in which backgrounds are "field-agnostic", allowing classes (such as Trajectory, Diffusion, or Distribution) to request that distinct fields types be computed by the background class. This resolves a long-standing issue where the Diffusion type's modularity with respect to the Trajectory type was proving to be somewhat elusive.
Also parametrizes background types via a general-purpose compile-time struct HyperParams, and started work on incorporating all the existing (hyper)parameters in the code into HyperParams, including the RK_INTEGRATOR_TYPE and the TRAJECTORY_TIME_FLOW.

Some updates also made to make the coordinates configurable via a Coordinates type, which opens a path to a future implementation of further coordinate-independence as well as unit-checking.

Also added a compile-time Debug and Release mode switch and started using these to condition some of the checks that can be skipped in release mode.