Source term update (#68)
This update ports the source term functionality implemented in the DSA fork back to the main SPECTRUM repository. Trajectories now have two additional variables, a path amplitude and a path weight, which are tracked throughout a simulation and factor into the final particle weight during binning. In particular, a TrajectoryParkerSource class was created to test this implementation through a new diffusive shock acceleration benchmark test. Source code was edited in several files, so please carefully review all the changes.

Other notable changes are:

 - The status messages from SimulationMaster every 1% of simulation work done have been improved. The previous implementation used the time spent integrating trajectories to approximate the time left in the simulation, but this overlooked MPI communication (and other worker tasks), and could yield a VERY bad estimate. The new estimate is more accurate and applicable to different simulation setups, and an additional "simulation efficiency" metric is reported for the user's benefit.

 - In preparation for the imminent `fields` update, all the benchmark tests have been updated and are now working correctly. Any update from now on should verify that these are still functional before being proposed (pushed). An automated pass/fail script that runs these tests, and others, will be generated as part of this upcoming update.

 - Including header SPECTRUM files was standarized to **always** use double quotes. There was a mix between quotes and angled braces, which was confusing since other non-standard libraries (e.g. eigen3) used angle braces as well.