This update, in flight, starts the process of integrating the new Fields type and configuration framework with MPI as well as the BATL-SPECTRUM software interface.
Server frontend/backend pattern is revised. A server frontend is replaced with a "Data" Background category (hiding the server-worker parallelism behind an abstraction at the user level) and the server backends simply called servers. The common server-worker MPI routines are contained in a class labeled ServerInterface.

The Block abstraction is simplified to a cleanly straightforward data structure that is consistent across all server types, and all interaction with the reader is moved into the server classes.