This update, in flight, comes close to completing the process of integrating the new Fields type and configuration framework with MPI as well as the BATL-SPECTRUM software interface.
In particular, this update cleans up MPI configuration in mpi_config.hh, and establishes a base server class that operates via virtual dispatch to a small number of specializations in each true server type. A HyperConfigure object is defined that splits up the older SimulationConfig, to allow some parameters to live in an isolated SimulationConfig object (an overdue revision). Revisions to NumericalDerivatives and TrajectoryBase establishes a trajectory-background interface via a revised CommonFields method. With these modifications, the DerivativeData information has a well-defined consistent pathway for all Background types and modes: starting with CommonFields, the initial definition of dmax via EvaluateDmax, to storage in NumericalDerivatives (which always occurs), to possible modification during numeric gradient computations, to propagation into Diffusion classes and any other object who needs this data. The possibility that NumericalDerivatives are not computed is resolved at compile time, and this information is exploited to pack memory and loosen memory bottlenecks. There is also partial completion of stencil interpolation using Field types, which (once completed) establishes a format-agnostic implementation of all data backgrounds.